Added image_scroller, design update
I changed the table to be less of an eyesore by adding whitespace and
collapsing the borders and giving the headers a nice grey background.
Further I reshuffled the search input boxes. Lastly I added
image_scroller which is incomplete but should be a framework we can work
on in the future.